separation of components + logging
